Handle revisions in play-by-play data
I noticed that sometimes the play-by-play csv files have revisions when an event was forgotten and it was added at a later time. This means the event has a higher EVENTNUM, but the PCTIMESTRING displays the actual time of the event. This commit changes the starters parser to handle such revisions, whereas it could not in the past.

I noticed one issue: When a player is subbed in and out during the same period and time on PCTIMESTRING and the first event was missed and included as a revision. This could be being subbed in before a free throw, the free throw was made, a timeout was called, and being subbed out. This is an incredibly rare event and the only game IDs I found it being an issue are
0029701045
0020500090
0022201040